scripts/scraper/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
feedparser>=6.0.0
python-dateutil>=2.8.0
pyyaml>=6.0.0
